FEATURE -- ENV -- COMPLETE

+++ utils/lib_len.c
@@ -6,7 +6,7 @@
 /*   By: rihoy <rihoy@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/26 16:22:05 by rihoy             #+#    #+#             */
-/*   Updated: 2024/03/07 18:13:14 by rihoy            ###   ########.fr       */
+/*   Updated: 2024/03/13 18:00:00 by rihoy            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ size_t	sent_len(char **sent)
 	return (i);
 }
 
-size_t	str_len(char *str)
+int	str_len(char *str)
 {
 	size_t	i;
